add block in css to card and hero and homecard to try fix display image
